Implement a new type for hosts files
The humble hosts file, along with zones and records, is one of the
three main domain objects this program needs to deal with.  So
representing these directly, rather than only as an implicit thing
used to construct a zone, makes things a bit clearer.
