+++ README.md
@@ -37,7 +37,7 @@ Development
 The project structure should hopefully be fairly straightforward.  The
 modules are:
 
-- `hosts`    - the parser for hosts files
+- `hosts`    - hosts files
 - `net_util` - shared utilities used by both the `main.rs` file and the `resolver` module
 - `protocol` - the DNS message types and serialisation / deserialisation logic
 - `resolver` - the resolution and caching logic
